Spark: block DROP TABLE when the user cannot delete the table data on HDFS (1.11.0)

+++ spark/v3.4/spark/src/main/java/org/apache/iceberg/spark/SparkCatalog.java
@@ -137,6 +137,7 @@ public class SparkCatalog extends BaseCatalog {
   private ViewCatalog asViewCatalog = null;
   private String[] defaultNamespace = null;
   private HadoopTables tables;
+  private Configuration hadoopConf;
 
   /**
    * Build an Iceberg {@link Catalog} to be used by this Spark catalog adapter.
@@ -400,10 +401,27 @@ public class SparkCatalog extends BaseCatalog {
 
   private boolean dropTableWithoutPurging(Identifier ident) {
     if (isPathIdentifier(ident)) {
-      return tables.dropTable(((PathIdentifier) ident).location(), false /* don't purge data */);
+      String location = ((PathIdentifier) ident).location();
+      DropTablePermissionValidator.validate(ident, location, hadoopConf);
+      return tables.dropTable(location, false /* don't purge data */);
     } else {
+      validateDeletePermission(ident);
       return icebergCatalog.dropTable(buildIdentifier(ident), false /* don't purge data */);
     }
+  }
+
+  private void validateDeletePermission(Identifier ident) {
+    org.apache.iceberg.Table table;
+    try {
+      table = icebergCatalog.loadTable(buildIdentifier(ident));
+    } catch (org.apache.iceberg.exceptions.NoSuchTableException
+        | org.apache.iceberg.exceptions.NotFoundException
+        | org.apache.iceberg.exceptions.RuntimeIOException e) {
+      // missing table or unreadable metadata: no data to protect, let the drop proceed
+      return;
+    }
+
+    DropTablePermissionValidator.validate(ident, table.location(), hadoopConf);
   }
 
   @Override
@@ -778,8 +796,8 @@ public class SparkCatalog extends BaseCatalog {
 
     this.catalogName = name;
     SparkSession sparkSession = SparkSession.active();
-    this.tables =
-        new HadoopTables(SparkUtil.hadoopConfCatalogOverrides(SparkSession.active(), name));
+    this.hadoopConf = SparkUtil.hadoopConfCatalogOverrides(sparkSession, name);
+    this.tables = new HadoopTables(hadoopConf);
     this.icebergCatalog =
         cacheEnabled
             ? CachingCatalog.wrap(catalog, cacheCaseSensitive, cacheExpirationIntervalMs)
